+++ wire-ios-data-model/Source/Utilis/EAR/EARStorage.swift
@@ -1,0 +1,58 @@
+//
+// Wire
+// Copyright (C) 2023 Wire Swiss GmbH
+//
+// This program is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with this program. If not, see http://www.gnu.org/licenses/.
+//
+
+import Foundation
+
+@objc
+public final class EARStorage: NSObject {
+
+    // MARK: - Properties
+
+    private let storage: PrivateUserDefaults<Key>
+
+    // MARK: - Types
+
+    private enum Key: String, DefaultsKey {
+        case enabledEAR = "com.wire.ear.enabled"
+    }
+
+    // MARK: - Life cycle
+
+    @objc
+    public init(
+        userID: UUID,
+        sharedUserDefaults: UserDefaults
+    ) {
+        storage = PrivateUserDefaults(
+            userID: userID,
+            storage: sharedUserDefaults
+        )
+
+        super.init()
+    }
+
+    // MARK: - Methods
+
+    public func earEnabled() -> Bool {
+        return storage.bool(forKey: .enabledEAR)
+    }
+
+    public func enableEAR(_ enabled: Bool) {
+        storage.set(enabled, forKey: .enabledEAR)
+    }
+}
