+++ wire-ios-notification-engine/WireNotificationEngineTests/MockEARServiceInterface.swift
@@ -16,7 +16,7 @@
 // along with this program. If not, see http://www.gnu.org/licenses/.
 //
 
-import Foundation
+import WireDataModel
 import LocalAuthentication
 
 public class MockEARServiceInterface: EARServiceInterface {
