+++ wire-ios-sync-engine/Source/UserSession/ZMUserSession.swift
@@ -282,10 +282,7 @@ public class ZMUserSession: NSObject {
             canPerformKeyMigration: true,
             sharedUserDefaults: sharedUserDefaults
         )
-        
-        // As we move the flag value from CoreData to UserDefaults, we set an initial value
-        self.earService.setInitialEARFlagValue(viewContext.encryptMessagesAtRest)
-        
+                
         self.lastEventIDRepository = LastEventIDRepository(
             userID: userId,
             sharedUserDefaults: sharedUserDefaults
@@ -293,6 +290,8 @@ public class ZMUserSession: NSObject {
 
         super.init()
 
+        // As we move the flag value from CoreData to UserDefaults, we set an initial value
+        self.earService.setInitialEARFlagValue(viewContext.encryptMessagesAtRest)
         self.earService.delegate = self
         appLockController.delegate = self
