+++ wire-ios-sync-engine/Tests/Source/Synchronization/EventProcessorTests.swift
@@ -215,10 +215,11 @@ class EventProcessorTests: MessagingTest {
         if isDatabaseLocked {
             let earService = EARService(
                 accountID: userIdentifier,
-                databaseContexts: [uiMOC, syncMOC]
-                earEnabledStorage: UserDefaults.random()
+                databaseContexts: [uiMOC, syncMOC],
+                sharedUserDefaults: UserDefaults.random()!
             )
-
+            earService.setInitialEARFlagValue(true)
+            
             try earService.enableEncryptionAtRest(
                 context: syncMOC,
                 skipMigration: true
